[refactor]: move args tests to `lib.rs`
