+++ cli/src/main.rs
@@ -33,57 +33,3 @@ async fn main() -> Result<()> {
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use iroha::is_colouring_supported;
-
-    use super::*;
-
-    #[test]
-    #[allow(clippy::bool_assert_comparison)] // for expressiveness
-    fn default_args() -> Result<()> {
-        let args = Args::try_parse_from(["test"])?;
-
-        assert_eq!(args.terminal_colors, is_colouring_supported());
-        assert_eq!(args.submit_genesis, false);
-
-        Ok(())
-    }
-
-    #[test]
-    #[allow(clippy::bool_assert_comparison)] // for expressiveness
-    fn terminal_colors_works_as_expected() -> Result<()> {
-        fn try_with(arg: &str) -> Result<bool> {
-            Ok(Args::try_parse_from(["test", arg])?.terminal_colors)
-        }
-
-        assert_eq!(
-            Args::try_parse_from(["test"])?.terminal_colors,
-            is_colouring_supported()
-        );
-        assert_eq!(try_with("--terminal-colors")?, true);
-        assert_eq!(try_with("--terminal-colors=false")?, false);
-        assert_eq!(try_with("--terminal-colors=true")?, true);
-        assert!(try_with("--terminal-colors=random").is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn user_provided_config_path_works() -> Result<()> {
-        let args = Args::try_parse_from(["test", "--config", "/home/custom/file.json"])?;
-
-        assert_eq!(args.config, Some(PathBuf::from("/home/custom/file.json")));
-
-        Ok(())
-    }
-
-    #[test]
-    fn user_can_provide_any_extension() {
-        let _args = Args::try_parse_from(["test", "--config", "file.toml.but.not"])
-            .expect("should allow doing this as well");
-    }
-}
